requirements.txt
Flask==2.1.1
psycopg2-binary==2.9.1
redis==3.5.3
